requirements.txt
homeassistant-api>=4.0.0
pyyaml>=6.0
requests>=2.28.1
aiohttp>=3.8.3
jinja2>=3.1.2
pytest>=7.2.0
pytest-asyncio>=0.20.0
pydantic>=1.10.5